+++ geolocator_platform_interface/lib/src/enums/location_permission.dart
@@ -4,8 +4,8 @@ enum LocationPermission {
   /// to request permission using the `Geolocator.requestPermission()` method.
   denied,
 
-  /// Permission to access the device's location is permenantly denied. When
-  /// requestiong permissions the permission dialog will not been shown until
+  /// Permission to access the device's location is permanently denied. When
+  /// requesting permissions the permission dialog will not be shown until
   /// the user updates the permission in the App settings.
   deniedForever,
 
@@ -17,8 +17,8 @@ enum LocationPermission {
   /// App is running in the background.
   always,
 
-  /// Permission status is cannot be determined. This permission is only
+  /// Permission status cannot be determined. This permission is only
   /// returned by the `Geolocator.checkPermission()` method on the web platform
-  /// for browsers that do not implement the Permission API (see https://developer.mozilla.org/en-US/docs/Web/API/Permissions_API).
+  /// for browsers that did not implement the Permission API (see https://developer.mozilla.org/en-US/docs/Web/API/Permissions_API).
   unableToDetermine
 }