+++ geolocator_web/example/lib/main.dart
@@ -52,7 +52,7 @@ class _GeolocatorWidgetState extends State<GeolocatorWidget> {
     );
 
     return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.background,
+      backgroundColor: Theme.of(context).colorScheme.surface,
       body: ListView.builder(
         itemCount: _positionItems.length,
         itemBuilder: (context, index) {