+++ geolocator/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 12.0.0
+
+- **BREAKING CHANGE:** Updates dependency on geolocator_web to version [4.0.0](https://pub.dev/packages/geolocator_web/changelog).
+
+## 11.1.0
+
+- Expose `AndroidPosition` from geolocator.
+
 ## 11.0.0
 
 - **BREAKING CHANGE:** Updates dependency on geolocator_web to version [3.0.0](https://pub.dev/packages/geolocator_web/changelog).